GW is almost able to connect again
svn path=/trunk/KDE/kdenetwork/kopete/; revision=736307

+++ protocols/groupwise/gwbytestream.h
@@ -57,7 +57,7 @@ private slots:
 	void slotConnected ();
 	void slotConnectionClosed ();
 	void slotReadyRead ();
-	void slotBytesWritten ( int );
+	void slotBytesWritten ( qint64 );
 	void slotError ( int );
 
 private:
